Removed default.js
