+++ index.js
@@ -1,6 +1,26 @@
-const defaults = require('./default')
+function toStar(length) {
+  let stars = ''
+  for (let i = 0; i < length; i++) {
+    stars += '*'
+  }
+  return stars
+}
 
-console.log(defaults)
+function getDefaultSwears() {
+  return new Set([
+    'fuck', 'shit', 'bitch', 'nigger', 'cock', 'pussy', 'pussies',
+    'kike', 'dyke', 'kyke', 'gook', 'wetback', 'penis', 'ass',
+    'dick', 'kraut', 'fag', 'cunt', 'twat', 'whore', 'douche'
+  ])
+}
+
+function getCensors() {
+  const censors = {}
+  getDefaultSwears.forEach(swear => {
+    censors[swear] = toStar(swear.length)
+  })
+  return censors
+}
 
 function parseMessage(message, delimiters) {
   let gathered = ''
